refactor: 도서 캐러셀에서 BookCard 분리
