+++ src/components/carousel/BooksInfinityCarousel/_mixins.scss
@@ -1,0 +1,26 @@
+@use './variables' as variables;
+@use '@/styles/mixins.scss' as mixins;
+
+@mixin fullWidthCard {
+  display: flex;
+  position: relative;
+  align-items: center;
+  gap: variables.$full-width-card__gap-desktop;
+  padding: 3rem;
+  height: 22rem;
+  @include mixins.full-width-card;
+
+  @media screen and (max-width: 768px) {
+    height: 20rem;
+  }
+
+  @media screen and (max-width: 425px) {
+    padding: 2rem;
+    height: 15rem;
+  }
+
+  @media screen and (max-width: 375px) {
+    gap: variables.$full-width-card__gap-mobile;
+    padding: 1.5rem;
+  }
+}
