+++ src/components/carousel/BooksInfinityCarousel/_variables.scss
@@ -1,0 +1,2 @@
+$full-width-card__gap-mobile: 1rem;
+$full-width-card__gap-desktop: 1.5rem;
