+++ src/components/carousel/BooksInfinityCarousel/components/BookCard/index.module.scss
@@ -1,0 +1,89 @@
+@use '../../variables' as variables;
+@use '../../mixins' as carouselMixins;
+@use '@/styles/mixins.scss' as mixins;
+
+$img-width-desktop: 10rem;
+$img-width-mobile: 8rem;
+$img-width-small-mobile: 7rem;
+
+.imgWrapper {
+  position: relative;
+
+  box-shadow: 0 0 1rem 0 rgba(0, 0, 0, 0.5);
+  aspect-ratio: 2/3;
+  width: $img-width-desktop;
+
+  @media screen and (max-width: 425px) {
+    width: $img-width-mobile;
+  }
+
+  @media screen and (max-width: 375px) {
+    width: $img-width-small-mobile;
+  }
+}
+
+.fullWidthCard {
+  @include carouselMixins.fullWidthCard;
+}
+
+.bookInfo {
+  display: flex;
+  position: relative;
+  flex-direction: column;
+  justify-content: center;
+  gap: 1rem;
+  z-index: 1;
+
+  padding-right: 1.5rem;
+
+  width: calc(100% - $img-width-desktop - variables.$full-width-card__gap-desktop);
+
+  h3,
+  li:not(:last-child) {
+    width: 100%;
+
+    @include mixins.ellipsis;
+  }
+  li {
+    font-size: 1.2rem;
+  }
+
+  @media screen and (max-width: 425px) {
+    gap: 0.8rem;
+    width: calc(100% - $img-width-mobile - variables.$full-width-card__gap-desktop);
+  }
+
+  @media screen and (max-width: 375px) {
+    width: calc(100% - $img-width-small-mobile - variables.$full-width-card__gap-mobile);
+  }
+}
+
+.bookTitle {
+  font-weight: var(--font-weight-semi-bold);
+  font-size: var(--font-size-basic);
+
+  @media screen and (max-width: 320px) {
+    font-size: var(--font-size-small);
+  }
+}
+
+.bookAuthorAndPublisher {
+  color: var(--color-gray500);
+}
+
+.bookContent {
+  display: -webkit-box;
+  margin-top: 0.5rem;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  -webkit-line-clamp: 3;
+  -webkit-box-orient: vertical;
+
+  @media screen and (max-width: 768px) {
+    -webkit-line-clamp: 2;
+  }
+  @media screen and (max-width: 425px) {
+    -webkit-line-clamp: 1;
+    font-size: var(--font-size-small);
+  }
+}
