+++ src/components/carousel/BooksInfinityCarousel/index.module.scss
@@ -1,10 +1,5 @@
 @use '@/styles/mixins.scss' as mixins;
-
-$full-width-card__gap-mobile: 1rem;
-$full-width-card__gap-desktop: 1.5rem;
-$img-width-desktop: 10rem;
-$img-width-mobile: 8rem;
-$img-width-small-mobile: 7rem;
+@use './mixins' as carouselMixins;
 
 .booksCarouselWrapper {
   display: flex;
@@ -19,105 +14,7 @@ $img-width-small-mobile: 7rem;
 }
 
 .fullWidthCard {
-  display: flex;
-  position: relative;
-  align-items: center;
-  gap: $full-width-card__gap-desktop;
-  padding: 3rem;
-  height: 22rem;
-  @include mixins.full-width-card;
-
-  @media screen and (max-width: 768px) {
-    height: 20rem;
-  }
-
-  @media screen and (max-width: 425px) {
-    padding: 2rem;
-    height: 15rem;
-  }
-
-  @media screen and (max-width: 375px) {
-    gap: $full-width-card__gap-mobile;
-    padding: 1.5rem;
-  }
-}
-
-.imgWrapper {
-  position: relative;
-
-  box-shadow: 0 0 1rem 0 rgba(0, 0, 0, 0.5);
-  aspect-ratio: 2/3;
-  width: $img-width-desktop;
-
-  @media screen and (max-width: 425px) {
-    width: $img-width-mobile;
-  }
-
-  @media screen and (max-width: 375px) {
-    width: $img-width-small-mobile;
-  }
-}
-
-.bookInfo {
-  display: flex;
-  position: relative;
-  flex-direction: column;
-  justify-content: center;
-  gap: 1rem;
-  z-index: 1;
-
-  padding-right: 1.5rem;
-
-  width: calc(100% - $img-width-desktop - $full-width-card__gap-desktop);
-
-  h3,
-  li:not(:last-child) {
-    width: 100%;
-
-    @include mixins.ellipsis;
-  }
-  li {
-    font-size: 1.2rem;
-  }
-
-  @media screen and (max-width: 425px) {
-    gap: 0.8rem;
-    width: calc(100% - $img-width-mobile - $full-width-card__gap-desktop);
-  }
-
-  @media screen and (max-width: 375px) {
-    width: calc(100% - $img-width-small-mobile - $full-width-card__gap-mobile);
-  }
-}
-
-.bookTitle {
-  font-weight: var(--font-weight-semi-bold);
-  font-size: var(--font-size-basic);
-
-  @media screen and (max-width: 320px) {
-    font-size: var(--font-size-small);
-  }
-}
-
-.bookAuthorAndPublisher {
-  color: var(--color-gray500);
-}
-
-.bookContent {
-  display: -webkit-box;
-  margin-top: 0.5rem;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  -webkit-line-clamp: 3;
-  -webkit-box-orient: vertical;
-
-  @media screen and (max-width: 768px) {
-    -webkit-line-clamp: 2;
-  }
-  @media screen and (max-width: 425px) {
-    -webkit-line-clamp: 1;
-    font-size: var(--font-size-small);
-  }
+  @include carouselMixins.fullWidthCard;
 }
 
 .cardSkeleton {
